third_party/WebKit/LayoutTests/fast/dom/shadow/content-element-includer-expected.txt

This tests the correctness of includers of forwarded children.
Note that this test needs internals object thus cannot run outside DRT.
PASS includerFor(childOfElementWithoutShadow) is null
PASS includerFor(childOfElementWithShadow) is null
PASS includerFor(childOfElementWithShadowContent) is shadowContentOfElementWithShadowContent
PASS includerFor(movingChild) is null
PASS includerFor(movingChild) is null
PASS includerFor(movingChild) is anotherShadowContentOfElementWithShadowContent
PASS successfullyParsed is true

TEST COMPLETE
